Split section handling of ConfigForm into new ConfigSectionForm
